Modified code to match string in phrase
In original project fork, the user must specifically state the color to change the background to. For example, user must say "red", "blue", "green", etc. exactly for the color change to process.

I changed how the recognition.onresult function sets the background color by running through a loop of the colors array, and seeking a match to the colors array using .includes() against the returned text from voice input.

With this change, the user can now say a phrase such as "I like the color red", and the function will detect the "red" and change the background accordingly.